// src/types/customer-details.ts
export type CustomerDetails = {
  id: string
  name: string
  email: string
  phone: string
  altPhone1: string
  altPhone2: string
  address1: string
  address2: string
  dob: string
  status: string
  maritalStatus: string
  anniversaryDate: string
  divorced: string
  notes: string
  banned: boolean
  noCall: boolean
  inactive: boolean
  optOutEcm: boolean
}